Added Inventory System and Item Pickup System
Created an inventory system that will allow players to pick up objects that are interactable, and add them to their inventory. There is a pickup tooltip that will be displayed for a few seconds on the screen that tells you what Item you just picked up. Also touched up black screen and eye opening system, and added a death zone to the testing level.