Added functionality to player inventory, and began work on implementing a Key System.
I added a RemoveItem and RemoveItemCheck set of functions to the InventoryObject class that will allow you to call on these from other scripts to make objects be able to 'take' an item from your inventory away from you. There are checks so that it will only take the item if you have at least 1. The check function is so that you can have conditional statements that will tell you 'you don't have the required item' if the check fails. See RayInteract script, Test_door and TestKey for functionality example.