Added Flashlight Pickup and System
Added a flashlight item that can be picked up in the gameworld. Once picked up it attaches to the player camera, and can be turned on and off as long as the item is in the player inventory. I also made the flashlight move upward when crouched to mimic the effect of raising a flashlight higher up if you were in a crouching position.