Added ending to Fear of Heights Level
Added Plane Crash model and fire effect to Fear of Heights level, as well as adding a ramp on the end of the spike building that leads to the roof. From there I added a platform where you have to take a 'leap of faith' forward, but if you go too far to either side you will fall on spikes. At the end of the path between the spikes is a bed.