Reworked footstep audio
The original footstep audio sounded uneven when walking and it was because the clips weren't all properly lined up. I went back and redid the same sounds again, but this time made sure to  do a better job of lining them up completely the same. Footsteps now sound much better. Also recorded demo for this week.